requirements.txt
numpy
matplotlib
torch
torchvision
pillow